Create Protobatch.js
protobatcher - the first iteration of batching. this manager script will require the existance of grow.js, weak.js, and hack.js that initiate grow();, hack();, and weaken(); with target parameters. this script requires a target server parameter. this script fires grow, hack, and weaken scripts all at the same time, instead of one at a time like the early-hack-template.js does in bitburner. the default % of money to be stolen from the server per run is 2%. you can change this, but this script does not incorporate the logic required to stop itself if the home server does not have the RAM needed to run all the required threads. this loop uses the Hack, Grow, Weaken logic, or HGW, which is ram inefficient, but is simple, since running them all at once makes the end timings line up properly.