Update Protobatch.js
removed unused functions, updated from HGW to HWGW, the superior ram efficiency model, and updated log loop to include all the time variables so the log is always producing pertinent data. also included logic to count down a start time for grow and hack, then switch to the normal countdown. the spacer is 30 milliseconds, so that won't ever show in the countdown as useable data since the log refreshes once per second at the quickest (1000 milliseconds). this should be the last update to this script, as this will be the basis for the actual batching script i make separately.